TOML: Move to encoding dir (#435)
